refactor: extract useBuilderLayers, useBuilderLadders, useBuilderMap from Builder.vue
